Extract operator to top-level class
